+++ src/main/java/com/idrsolutions/microservice/JPedalServlet.java
@@ -23,18 +23,20 @@ package com.idrsolutions.microservice;
 import com.idrsolutions.image.utility.SupportedFormats;
 import com.idrsolutions.microservice.db.DBHandler;
 import com.idrsolutions.microservice.storage.Storage;
+import com.idrsolutions.microservice.utils.ConversionTracker;
 import com.idrsolutions.microservice.utils.DefaultFileServlet;
 import com.idrsolutions.microservice.utils.LibreOfficeHelper;
 import com.idrsolutions.microservice.utils.SettingsValidator;
 import com.idrsolutions.microservice.utils.ZipHelper;
+import org.jpedal.PdfDecoderServer;
 import org.jpedal.examples.images.ConvertPagesToImages;
 import org.jpedal.examples.images.ExtractClippedImages;
 import org.jpedal.examples.images.ExtractImages;
 import org.jpedal.examples.text.ExtractStructuredText;
 import org.jpedal.examples.text.ExtractTextAsWordlist;
 import org.jpedal.examples.text.ExtractTextInRectangle;
+import org.jpedal.exception.PdfException;
 import org.jpedal.external.ErrorTracker;
-import org.jpedal.io.DefaultErrorTracker;
 import org.w3c.dom.Document;
 
 import javax.json.stream.JsonParsingException;
@@ -144,6 +146,29 @@ public class JPedalServlet extends BaseServlet {
         //Makes the directory for the output file
         new File(outputDirStr + fileSeparator + fileNameWithoutExt).mkdirs();
 
+        final int pageCount;
+        try {
+            final PdfDecoderServer decoder = new PdfDecoderServer(false);
+            decoder.openPdfFile(inputFile.getAbsolutePath());
+
+            decoder.setEncryptionPassword(conversionParams.getOrDefault("org.jpedal.pdf2html.password", ""));
+
+            if (decoder.isEncrypted() && !decoder.isPasswordSupplied()) {
+                LOG.log(Level.SEVERE, "Invalid Password");
+                DBHandler.getInstance().setError(uuid, 1070, "Invalid password supplied.");
+                return;
+            }
+
+            pageCount = decoder.getPageCount();
+            DBHandler.getInstance().setCustomValue(uuid, "pageCount", String.valueOf(pageCount));
+            DBHandler.getInstance().setCustomValue(uuid, "pagesConverted", "0");
+            decoder.closePdfFile();
+            decoder.dispose();
+        } catch (final PdfException e) {
+            LOG.log(Level.SEVERE, "Invalid PDF", e);
+            DBHandler.getInstance().setError(uuid, 1060, "Invalid PDF");
+            return;
+        }
         DBHandler.getInstance().setState(uuid,"processing");
 
         try {
@@ -158,7 +183,7 @@ public class JPedalServlet extends BaseServlet {
 
             final long maxDuration = Long.parseLong(properties.getProperty(BaseServletContextListener.KEY_PROPERTY_MAX_CONVERSION_DURATION));
 
-            convertPDF(mode, userPdfFilePath, outputDirStr, fileNameWithoutExt, conversionParams, new DurationTracker(uuid, maxDuration));
+            convertPDF(mode, userPdfFilePath, outputDirStr, fileNameWithoutExt, conversionParams, new ConversionTracker(uuid, maxDuration));
 
             if ("1230".equals(DBHandler.getInstance().getStatus(uuid).get("errorCode"))) {
                 final String message = String.format("Conversion %s exceeded max duration of %dms", uuid, maxDuration);
@@ -327,29 +352,4 @@ public class JPedalServlet extends BaseServlet {
             super(msg);
         }
     }
-
-    private static class DurationTracker extends DefaultErrorTracker {
-
-        private final String uuid;
-        private final long startTime;
-        private final long maxDuration;
-
-        public DurationTracker(final String uuid, final long maxDuration) {
-            this.uuid = uuid;
-            this.maxDuration = maxDuration;
-            startTime = System.currentTimeMillis();
-        }
-
-        @Override
-        public boolean checkForExitRequest(int dataPointer, int streamSize) {
-            if (System.currentTimeMillis() - startTime > maxDuration) {
-                DBHandler.getInstance().setError(uuid, 1230, "Conversion exceeded max duration of " + maxDuration + "ms");
-                return true;
-            }
-
-            return false;
-        }
-
-    }
-
 }
